requirements.txt
# =============================================================================
# GNN-IDS: Graph Neural Network for Intrusion Detection System
# Requirements file for Python dependencies
# =============================================================================

# Core Deep Learning Framework
torch>=2.0.0,<2.3.0
torchvision>=0.15.0

# PyTorch Geometric and dependencies
torch-geometric>=2.3.0,<2.5.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0
torch-spline-conv>=1.2.0

# Data Processing & Scientific Computing
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<1.27.0
scikit-learn>=1.3.0,<1.5.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0,<3.9.0
seaborn>=0.12.0,<0.14.0
plotly>=5.14.0,<6.0.0

# Machine Learning Metrics & Monitoring
tensorboard>=2.13.0,<3.0.0

# Utilities
tqdm>=4.65.0,<5.0.0
pyyaml>=6.0,<7.0
python-dotenv>=1.0.0

# Web Framework for Demo (optional)
streamlit>=1.28.0,<2.0.0
gradio>=3.50.0,<5.0.0

# Dataset handling
wget>=3.2
requests>=2.31.0,<3.0.0

# Network & Graph Analysis
networkx>=3.1,<4.0

# Development Dependencies (optional, install with: pip install -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.1.0
# mypy>=1.5.0
# ipython>=8.14.0
# jupyter>=1.0.0
# notebook>=7.0.0

# Additional useful packages
colorama>=0.4.6  # Colored terminal output

